+++ src/main/java/ru/services/MessegeService.java
@@ -19,12 +19,12 @@ public class MessegeService {
     }
 
     @Transactional
-    public void deleteMassageById(Long id) {
+    public void deleteMassageById(Integer id) {
         messageRepostory.deleteById(id);
     }
 
     @Transactional
-    public List<Message> getMessageByDialogId(Long id) {
+    public List<Message> getMessageByDialogId(Integer id) {
         return messageRepostory.findByDialogIdOrderByDate(id);
     }
 }
